+++ contracts/test-artifacts/MintableERC20.sol
@@ -12,7 +12,7 @@ contract MintableERC20 is
     ERC20Detailed(_name, _symbol, _decimals)
   {}
 
-  function mint(address _account, uint256 _amount) public returns (bool) {
+  function mint(address _account, uint _amount) public returns (bool) {
     // Stop excessive amounts so we don't overflow
     require(_amount < 1000000000000000000000000000000, "LOWER_AMOUNT");
     _mint(_account, _amount);
